# api/docs/swagger.yaml
definitions:
  http.Empty:
    type: object
  order_service.Order:
    properties:
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      productID:
        type: string
      totalSum:
        type: integer
      userID:
        type: string
    type: object
  user_service.User:
    properties:
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
info:
  contact: {}
  description: This is api gateway
paths:
  /order:
    post:
      consumes:
      - application/json
      description: Create Order
      operationId: create_order
      parameters:
      - description: CreateOrderRequest
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order_service.Order'
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            $ref: '#/definitions/order_service.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/order_service.Order'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/order_service.Order'
      summary: Create Order
      tags:
      - Order
  /order/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Order
      operationId: delete_order
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Empty'
      summary: Delete Order
      tags:
      - Order
    get:
      consumes:
      - application/json
      description: Get By ID Order
      operationId: get_order_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            $ref: '#/definitions/order_service.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/order_service.Order'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/order_service.Order'
      summary: Get By ID Order
      tags:
      - Order
  /user:
    post:
      consumes:
      - application/json
      description: Create User
      operationId: create_user
      parameters:
      - description: CreateUserRequest
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user_service.User'
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            $ref: '#/definitions/user_service.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user_service.User'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/user_service.User'
      summary: Create User
      tags:
      - User
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User
      operationId: delete_user
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete User
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get By ID User
      operationId: get_user_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            $ref: '#/definitions/user_service.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user_service.User'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/user_service.User'
      summary: Get By ID User
      tags:
      - User
swagger: "2.0"